test(e2e): refactor memory tests onto shared fixtures and split helpers/conftest
Replace the ad-hoc, duplicated memory-test tooling with reusable primitives and
organize the growing helper/conftest files by concern.

Primitives (any memory-capable adapter test can build on these):
- MemoryProbe via the `memory` fixture: marker() (meaningful + timestamped +
  random so the LLM keeps it and reruns can't collide), wait() (polls a scope,
  tracks matches, raises on subject scope without subject_id), archive-on-teardown.
- send_and_wait_for_reply: the standard trigger-and-wait helper.
Both encode the lessons from debugging so the same mistakes can't recur.

Structure:
- helpers.py -> helpers/ package by concern: log, messaging, agent, memory
  (__init__ re-exports the public API; call sites unchanged).
- conftest.py fixtures -> fixtures/ plugin modules (clients, rooms, memory)
  loaded via pytest_plugins; conftest keeps settings, markers, and the
  collection hook.

Behavior:
- e2e_fresh_room_allocator makes a clean room per test (no inherited history)
  and, on teardown, removes the agent from those rooms so they don't count
  against the 10-room cap (no chat-delete API exists); opt out with --no-clean.
- agno + langgraph memory tests use the shared primitives and fresh rooms.
- Memory scope guidance points to band_get_participants for an in-room subject's
  id (band_lookup_peers excludes in-room peers); fix stray backtick.
