+++ tests/e2e/adapters/test_agno_memory.py
@@ -7,9 +7,10 @@ passing test validates both the memory tools and the prompt-injection feature
 (the Agno adapter appends ``MEMORY_SECTION`` to the agent's system prompt when
 the memory capability is enabled).
 
-Each remembered fact carries a per-run UUID marker so it is identifiable on the
-live platform; the created memories are archived in teardown unless ``--no-clean``
-(or ``BAND_TEST_NO_CLEAN``) is set.
+Memory plumbing (unique markers, polling, teardown cleanup) comes from the shared
+``memory`` fixture (``MemoryProbe``); the trigger-and-wait flow from
+``send_and_wait_for_reply``. New memory tests should reuse those rather than
+re-implementing them.
 
 Run with:
     E2E_TESTS_ENABLED=true uv run pytest tests/e2e/adapters/test_agno_memory.py -v -s --no-cov
@@ -17,18 +18,13 @@ Run with:
 
 from __future__ import annotations
 
-import asyncio
-import contextlib
 from collections.abc import AsyncGenerator
-from typing import Any
-from uuid import uuid4
 
 import pytest
 from band_rest import AsyncRestClient
 
 from band import Agent
-from band.core.types import AdapterFeatures, Capability
-from tests.conftest_integration import is_no_clean_mode
+from band.core.types import AdapterFeatures, Capability, Emit
 from tests.e2e.conftest import (
     E2ESettings,
     RoomAllocator,
@@ -36,10 +32,10 @@ from tests.e2e.conftest import (
     requires_openai,
 )
 from tests.e2e.helpers import (
+    MemoryProbe,
     TrackingWebSocketClient,
-    listening_for_agent_responses,
     running_agent,
-    send_trigger_message,
+    send_and_wait_for_reply,
 )
 
 # Deliberately generic — no mention of scope/system/type/segment, so the agent
@@ -54,9 +50,11 @@ SECRETARY_INSTRUCTIONS = (
 
 @pytest.fixture
 async def agno_memory_room(
-    e2e_room_allocator: RoomAllocator,
+    e2e_fresh_room_allocator: RoomAllocator,
 ) -> tuple[str, str, str]:
-    return await e2e_room_allocator("agno-memory")
+    # A fresh room per test: the memory agent must not inherit unrelated history
+    # from reused rooms, which derails small models and pollutes the scope check.
+    return await e2e_fresh_room_allocator("agno-memory")
 
 
 @pytest.fixture
@@ -78,9 +76,15 @@ async def running_agno_memory_agent(
         model=OpenAIChat(id=e2e_config.e2e_llm_model),
         instructions=SECRETARY_INSTRUCTIONS,
     )
+    # Emit.EXECUTION posts the agent's tool_call/tool_result events to the room,
+    # so a failing run can be debugged by inspecting what the agent actually did
+    # (e.g. via band's REST context) instead of guessing.
     adapter = AgnoAdapter(
         agno_agent,
-        features=AdapterFeatures(capabilities={Capability.MEMORY}),
+        features=AdapterFeatures(
+            capabilities={Capability.MEMORY},
+            emit={Emit.EXECUTION},
+        ),
     )
 
     async with running_agent(
@@ -90,54 +94,6 @@ async def running_agno_memory_agent(
         config=e2e_config,
     ) as agent:
         yield agent
-
-
-@pytest.fixture
-async def archived_memory_ids(
-    e2e_session_client: AsyncRestClient,
-    request: pytest.FixtureRequest,
-) -> AsyncGenerator[list[str], None]:
-    """Collect memory IDs created by a test and archive them on teardown.
-
-    Tests append the IDs they verified. Archiving (hide but preserve) keeps the
-    live organization clean across runs. Honors ``--no-clean`` /
-    ``BAND_TEST_NO_CLEAN`` so data can be inspected after a run.
-    """
-    ids: list[str] = []
-    yield ids
-
-    if is_no_clean_mode(request):
-        return
-    for memory_id in ids:
-        with contextlib.suppress(Exception):
-            await e2e_session_client.agent_api_memories.archive_agent_memory(
-                id=memory_id
-            )
-
-
-async def _wait_for_memories(
-    client: AsyncRestClient,
-    marker: str,
-    *,
-    scope: str,
-    timeout: float,
-) -> list[Any]:
-    """Poll until active ``scope`` memories contain ``marker``; return the matches."""
-    deadline = asyncio.get_running_loop().time() + timeout
-    while asyncio.get_running_loop().time() < deadline:
-        response = await client.agent_api_memories.list_agent_memories(
-            page_size=50, status="active", scope=scope
-        )
-        matches = [
-            memory
-            for memory in response.data or []
-            if marker in (getattr(memory, "content", None) or "")
-        ]
-        if matches:
-            return matches
-        await asyncio.sleep(1)
-
-    pytest.fail(f"Expected {scope} memory containing {marker}")
 
 
 # loop_scope="session" runs the agent's background task on the test's event loop
@@ -150,36 +106,31 @@ async def test_agno_secretary_stores_organization_memory(
     e2e_config: E2ESettings,
     agno_memory_room: tuple[str, str, str],
     e2e_agent_info: tuple[str, str],
-    e2e_session_client: AsyncRestClient,
     e2e_user_client: AsyncRestClient,
     running_agno_memory_agent: Agent,
     ws_client: TrackingWebSocketClient,
-    archived_memory_ids: list[str],
+    memory: MemoryProbe,
 ) -> None:
     """A shared/company fact is stored as an organization-scoped memory."""
     chat_id, _user_id, _user_name = agno_memory_room
     agent_id, agent_name = e2e_agent_info
-    marker = f"AGNO_MEM_ORG_{uuid4().hex}"
+    marker = memory.marker("Q3LAUNCH")
     prompt = (
-        f"Remember this for the whole organization (so it can be shared everywhere): {marker} is the code name for our "
-        "Q3 launch."
+        "Remember this for the whole organization (so it can be shared "
+        f"everywhere): the code name for our Q3 launch is {marker}."
     )
 
-    async with listening_for_agent_responses(
-        ws_client, chat_id, timeout=e2e_config.e2e_timeout, raise_on_timeout=True
-    ) as wait_for_reply:
-        await send_trigger_message(
-            e2e_user_client, chat_id, prompt, agent_name, agent_id
-        )
-        await wait_for_reply()
-
-    matches = await _wait_for_memories(
-        e2e_session_client,
-        marker,
-        scope="organization",
+    await send_and_wait_for_reply(
+        ws_client,
+        e2e_user_client,
+        chat_id,
+        prompt,
+        agent_name,
+        agent_id,
         timeout=e2e_config.e2e_timeout,
     )
-    archived_memory_ids.extend(m.id for m in matches)
+
+    await memory.wait(marker, scope="organization")
 
 
 @pytest.mark.asyncio(loop_scope="session")
@@ -190,44 +141,39 @@ async def test_agno_secretary_stores_subject_memory(
     e2e_config: E2ESettings,
     agno_memory_room: tuple[str, str, str],
     e2e_agent_info: tuple[str, str],
-    e2e_session_client: AsyncRestClient,
     e2e_user_client: AsyncRestClient,
     running_agno_memory_agent: Agent,
     ws_client: TrackingWebSocketClient,
-    archived_memory_ids: list[str],
+    memory: MemoryProbe,
 ) -> None:
     """A personal fact is stored as a subject-scoped memory linked to the user.
 
     The agent is only told the fact is "about me specifically" — it must infer
-    subject scope and resolve the user's subject_id (via band_lookup_peers / the
-    participant list) from the injected memory-scope guidance.
+    subject scope and resolve the user's subject_id (via band_get_participants /
+    band_lookup_peers) from the injected memory-scope guidance.
     """
     chat_id, user_id, _user_name = agno_memory_room
     agent_id, agent_name = e2e_agent_info
-    marker = f"AGNO_MEM_SUBJ_{uuid4().hex}"
+    marker = memory.marker("BADGE")
     prompt = (
         "Remember this about me personally so you recall it whenever we talk: "
-        f"{marker} — I prefer espresso over drip coffee. Save it as being about "
-        "me specifically."
+        f"my employee badge number is {marker}. Save it as being about me "
+        "specifically."
     )
 
-    async with listening_for_agent_responses(
-        ws_client, chat_id, timeout=e2e_config.e2e_timeout, raise_on_timeout=True
-    ) as wait_for_reply:
-        await send_trigger_message(
-            e2e_user_client, chat_id, prompt, agent_name, agent_id
-        )
-        await wait_for_reply()
-
-    matches = await _wait_for_memories(
-        e2e_session_client,
-        marker,
-        scope="subject",
+    await send_and_wait_for_reply(
+        ws_client,
+        e2e_user_client,
+        chat_id,
+        prompt,
+        agent_name,
+        agent_id,
         timeout=e2e_config.e2e_timeout,
     )
+
+    matches = await memory.wait(marker, scope="subject", subject_id=user_id)
     assert any(getattr(m, "subject_id", None) == user_id for m in matches), (
         f"Expected a subject memory containing {marker} linked to subject "
         f"{user_id}, but matched subjects were "
         f"{[getattr(m, 'subject_id', None) for m in matches]}."
     )
-    archived_memory_ids.extend(m.id for m in matches)
